refactor: move access-token and password-reset-token tests to authentication
