+++ tests/modules/authentication/test_access_token_service.py
@@ -9,7 +9,7 @@ from modules.account.types import (
 )
 from modules.authentication.authentication_service import AuthenticationService
 from modules.authentication.types import CreateOTPParams
-from tests.modules.access_token.base_test_access_token import BaseTestAccessToken
+from tests.modules.authentication.base_test_access_token import BaseTestAccessToken
 
 
 class TestAuthenticationService(BaseTestAccessToken):
